aqa_python/automation_playwright/requirements.txt
httpx==0.28.1
pytest==8.3.5
python-dotenv==1.1.0
loguru==0.7.3
playwright==1.52.0